move handshake code to handshake.go
